% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atdb_get_data.R
\name{atdb_get_data}
\alias{atdb_get_data}
\title{Get data from SIPRI Arms Transfer Database (ATDB)}
\usage{
atdb_get_data(
  verbose = FALSE,
  start_year,
  end_year,
  footnotes = FALSE,
  cache = TRUE
)
}
\arguments{
\item{verbose}{Logical. If TRUE, prints additional details about the query
process. Default is FALSE.}

\item{start_year}{Numeric. Start of date range to query.}

\item{end_year}{Numeric. End of date range to query.}

\item{cache}{Logical. If TRUE, caches the response and processed
data to avoid re-querying. Default is TRUE.}
}
\value{
A tibble containing the requested data.
}
\description{
The \code{atdb_get_data} function retrieves arms transfer data from the
SIPRI arms transfer database (ATDB).
}
