% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atdb_request.R
\name{atdb_perform_request}
\alias{atdb_perform_request}
\title{Perform a request and retrieve response}
\usage{
atdb_perform_request(req)
}
\arguments{
\item{req}{An \code{httr2} request object created by \code{atdb_build_request}.}
}
\value{
An \code{httr2} response object containing the server's response.
}
\description{
Executes an HTTP request and retrieves the response.
}
\keyword{internal}
