% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atdb_pretty_cols}
\alias{atdb_pretty_cols}
\title{atdb_pretty_cols}
\format{
\code{column_names}
A dataframe with 17 rows and two columns:
\describe{
\item{to}{tidy columns}
\item{from}{original column names}
}
}
\usage{
atdb_pretty_cols
}
\description{
A data.frame with a matched list of tidy and
untidy column names for the results.
}
\keyword{datasets}
