% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atdb_request.R
\name{sipri_build_request}
\alias{sipri_build_request}
\title{Construct SIPRI ATDB Request}
\usage{
sipri_build_request(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical. If TRUE, prints detailed information about the request.}

\item{start_year}{Numeric. Start of the date range for querying data.}

\item{end_year}{Numeric. End of the date range for querying data.}
}
\value{
An \code{httr2} request object.
}
\description{
Prepares an HTTP POST request to the SIPRI Arms Transfer Database (ATDB) API.
}
\keyword{internal}
